decimal issue fix

+++ src/main/java/com/asg/finance/dto/BankPaymentGLDetailResponse.java
@@ -1,6 +1,7 @@
 package com.asg.finance.dto;
 
 import lombok.Data;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -16,13 +17,13 @@ public class BankPaymentGLDetailResponse {
     private LovGetListDto companyDet;
     private Long glPoid;
     private LovGetListDto glDet;
-    private Double drAmt;
-    private Double crAmt;
+    private BigDecimal drAmt;
+    private BigDecimal crAmt;
     private Long taxPoid;
     private LovGetListDto taxDet;
-    private Double taxPercentage;
-    private Double taxAmount;
-    private Double totalAmount;
+    private BigDecimal taxPercentage;
+    private BigDecimal taxAmount;
+    private BigDecimal totalAmount;
     private String partyInvNumber;
     private LocalDate partyInvDate;
     private String remarks;
